Now make gd-document class really "letter" format, and use better font.
When using geometry, it is important to tell the geometry package
about the paper size.

As for the better font, the class now uses the Libertine family,
which is a high-quality font available on any Linux distribution.
If for any reason you don't have a package in your Linux distribution's
repositories, you can pass the "defaultfont" option to the class,
in which case it will simply use LaTeX's default fonts.